+++ plugins/modules/lbvserver_servicegroup_binding.py
@@ -39,7 +39,7 @@ options:
   order:
     description:
       - Order number to be assigned to the service when it is bound to the lb vserver.
-    type: int
+    type: float
   servicegroupname:
     description:
       - The service group name bound to the selected load balancing virtual server.
@@ -54,7 +54,7 @@ options:
         greater weight. Defines the capacity of the service relative to the other
         services in the load balancing configuration. Determines the priority given
         to the service in load balancing decisions.
-    type: int
+    type: float
     default: 1
 extends_documentation_fragment: netscaler.adc.netscaler_adc
